Basic UI Layout
Created the basic UI layout and provided activities for each button. The
icons aren't labeled and all use the same icon, but each one is for a
different activity. UI doesn't respond to device resolution changes
well. It will probably need to be rewriten in the future, but this is
something to work off of for now.